+++ src/app/user/profile/page.tsx
@@ -1,10 +1,10 @@
 "use client";
 
 import Container from "@/components/container/Container";
-import Addresses from "@/components/forms/(user)/Addresses";
-import PaymentMethod from "@/components/forms/(user)/PaymentMethod";
-import PersonalInfo from "@/components/forms/(user)/(personal-info)/PersonalInfo";
-import Security from "@/components/forms/(user)/Security";
+import Addresses from "@/components/forms/user/Addresses";
+import PaymentMethod from "@/components/forms/user/PaymentMethod";
+import PersonalInfo from "@/components/forms/user/personal-info/PersonalInfo";
+import Security from "@/components/forms/user/Security";
 import { PROFILE_PAGE_TABS } from "@/lib/constants";
 import { useState } from "react";
